Error or success message displays on form submit. Gulp task to minify app.js and move bootstrap stylesheet and app.js to public folder.
