+++ package.json
@@ -14,5 +14,10 @@
   "homepage": "https://github.com/wsu-cs4350-group6/Project3",
   "dependencies": {
     "bower": "^1.3.12"
+  },
+  "devDependencies": {
+    "gulp": "^3.8.11",
+    "gulp-clean": "^0.3.1",
+    "gulp-uglify": "^1.1.0"
   }
 }
