+++ public/index.html
@@ -2,7 +2,7 @@
 <html>
 <head>
 	<title>Group 6 Login Form</title>
-	<link href="" rel="stylesheet">
+	<link href="stylesheets/bootstrap.min.css" rel="stylesheet">
 	<script src="https://ajax.googleapis.com/ajax/libs/jquery/2.1.3/jquery.min.js"></script>
 	<script src="javascripts/app.js"></script>
 </head>
