+++ public/javascripts/app.js
@@ -1,72 +1,1 @@
-(function(group6) {
-
-	group6(window.jQuery, window, document);
-
-}(function($, window, document) {
-
-	$(function() {
-
-		$("div#loginForm").append(
-			// Creating Form Div and Adding <h2> and <p> Paragraph Tag in it.
-			$("<h3/>").text("Sign In"),
-			$("<form/>", {
-				//action: '#',
-				method: 'post'
-			}).append(
-				// Create <form> Tag and Appending in HTML Div form1.
-				$("<input/>", {
-				type: 'text',
-				id: 'username',
-				name: 'username',
-				placeholder: 'username'
-				}), 
-				// Creating Input Element With Attribute.
-				$("<input/>", {
-				type: 'password',
-				id: 'password',
-				name: 'password',
-				placeholder: 'password'
-				}),
-				$("<br/>"), $("<input/>", {
-				type: 'submit',
-				id: 'submit',
-				value: 'Submit'
-				})
-			)		
-		);
-
-		var form = $("#loginForm");
-			form.on({
-
-				"click": function(e) {
-                    e.preventDefault();
-                    var username = $(this).find("[name='username']").val();
-                    var password = $(this).find("[name='password']").val();
-
-                    authenticate(username, password, "e618d316-8249-5d7a-8eac-8942f73192d7").done(function (data) {
-                        // Updates the UI based the ajax result
-
-                        $("#authorized").text("Authorized");
-
-                    })
-                    $("#authorized").text('');
-                }
-
-		});
-
-	});
-
-	function authenticate(username, password, accessKey) {
-		var key = accessKey;
-		var formData = {username:username, password:password};
-		return $.ajax({
-			url: "/authenticate",
-			type: "post",
-			headers: {
-				"Authorization": key
-			},
-			data: formData
-		});
-	}
-
-}));
+!function(e){e(window.jQuery,window,document)}(function(e){function n(n,t,a){var r=a,o={username:n,password:t};return e.ajax({url:"/authenticate",type:"post",headers:{Authorization:r},data:o})}function t(){return e.ajax({url:"/access",type:"get"})}e(function(){e("div#loginForm").append(e("<h3/>").text("Sign In"),e("<form/>",{method:"post"}).append(e("<input/>",{type:"text",id:"username",name:"username",placeholder:"username"}),e("<input/>",{type:"password",id:"password",name:"password",placeholder:"password"}),e("<br/>"),e("<input/>",{type:"submit",id:"submit",value:"Submit"})));var a=e("#loginForm");a.on({submit:function(a){a.preventDefault();{var r=e(this).find("[name='username']").val(),o=e(this).find("[name='password']").val();t().done(function(t){n(r,o,t.key).done(function(n){var t=e.parseJSON(n);e.each(t,function(n,t){e("#authorized").text("User profile url: "+t)})}).fail(function(n){var t=e.parseJSON(n.responseText);e.each(t,function(n,t){e("#authorized").text(t)})})})}}})})});
